requirements.txt
langgraph>=0.2.0
mistralai>=1.0.0
langfuse>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
